Revert Drive intent fetcher abstraction
Reverts the following commits to restore the pre-abstraction state:
- fb733ff75 Reuse buildResponse helper across Drive intent mapper tests
- c553b8f48 Record bridge and token Drive intent routes in ADR-0106 and ADR-0095
- 4564d1d79 Split Drive intent creation into bridge and token fetchers
