+++ .github/workflows/analyze-test.yaml
@@ -20,7 +20,7 @@ env:
     core/test/utils/html/editor_script/quoted_reply_enter_handler_script_browser_test.dart
     core/test/utils/html/mobile_email_responsive_layout_browser_test.dart
     core/test/utils/html/mobile_email_responsive_layout_side_effects_browser_test.dart
-    workplace/test/data/workplace_bridge_datasource_impl_web_test.dart
+    workplace/test/data/workplace_datasource_impl_web_test.dart
     workplace/test/presentation/extension/workplace_composer_attachment_extension_web_test.dart
 
 jobs:
