+++ workplace/lib/data/datasource_impl/workplace_datasource_impl.dart
@@ -1,23 +1,62 @@
+import 'dart:convert';
+
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 import 'package:workplace/data/model/workplace_exchange_token_response.dart';
 import '../model/workplace_exchange_token_request.dart';
 import '../datasource/workplace_datasource.dart';
-import '../mapper/workplace_intent_mapper.dart';
 import '../model/workplace_enums.dart';
+import '../model/workplace_intent_request.dart';
+import '../model/workplace_intent_response.dart';
+import '../bridge/cozy_bridge.dart';
 import '../workplace_dio.dart';
 import '../../domain/entity/workplace_intent.dart';
+import '../../domain/entity/workplace_intent_access_mode.dart';
 import '../../domain/entity/workplace_intent_config.dart';
 
-/// Token route: direct HTTP to Drive, authenticated with an exchanged bearer token.
 class WorkplaceDataSourceImpl implements WorkplaceDataSource {
   WorkplaceDataSourceImpl();
+
+  Map<String, dynamic> _asJsonMap(dynamic data) {
+    if (data is Map<String, dynamic>) return data;
+    if (data is String) return jsonDecode(data) as Map<String, dynamic>;
+    throw FormatException(
+      'Expected JSON object or string, got: ${data.runtimeType}',
+    );
+  }
 
   @override
   Future<WorkplaceIntent> createIntent({
     required Uri platformUrl,
-    required String accessToken,
+    required WorkplaceIntentAccessMode accessMode,
     required WorkplaceIntentConfig config,
   }) async {
+    final body = _buildIntentRequest(config);
+
+    return switch (accessMode) {
+      BridgeAccessMode() => _createIntentViaBridge(body),
+      BearerTokenAccessMode(:final accessToken) =>
+        _createIntentViaBearerToken(platformUrl, accessToken, body),
+    };
+  }
+
+  // No token: the container app already holds the stack session.
+  Future<WorkplaceIntent> _createIntentViaBridge(
+    Map<String, dynamic> body,
+  ) async {
+    final data = await CozyBridge.fetchJson(
+      method: 'POST',
+      path: '/intents',
+      body: body,
+    );
+    return parseIntentResponse(data);
+  }
+
+  Future<WorkplaceIntent> _createIntentViaBearerToken(
+    Uri platformUrl,
+    String accessToken,
+    Map<String, dynamic> body,
+  ) async {
     // Fail fast instead of sending a malformed Authorization header.
     if (accessToken.trim().isEmpty) {
       throw StateError('Access token is empty');
@@ -36,10 +75,52 @@ class WorkplaceDataSourceImpl implements WorkplaceDataSource {
       options: Options(
         headers: {'Authorization': 'Bearer $accessToken'},
       ),
-      data: buildIntentRequestBody(config),
+      data: body,
     );
     return parseIntentResponse(response.data);
   }
+
+  WorkplaceIntent parseIntentResponse(
+    dynamic data, {
+    bool requireHttps = kReleaseMode,
+  }) {
+    final parsed = WorkplaceIntentResponse.fromJson(_asJsonMap(data));
+    final services = parsed.data.attributes.services;
+    if (services.isEmpty) {
+      throw StateError('Drive response contains no services');
+    }
+
+    final href = services.first.href;
+    final intentUrl = Uri.parse(href);
+    if (requireHttps && intentUrl.scheme != 'https') {
+      throw ArgumentError('Intent URL must use HTTPS, got: $href');
+    }
+
+    return WorkplaceIntent(
+      intentId: parsed.data.id,
+      intentUrl: intentUrl,
+      client: parsed.data.attributes.client,
+    );
+  }
+
+  Map<String, dynamic> _buildIntentRequest(WorkplaceIntentConfig config) =>
+      WorkplaceIntentRequest(
+    data: WorkplaceIntentDataRequest(
+      type: WorkplaceDataRequestType.intents,
+      attributes: WorkplaceIntentAttributesRequest(
+        action: WorkplaceAction.pick,
+        type: WorkplaceDocType.files,
+        permissions: [WorkplacePermission.get],
+        data: WorkplaceFilePickerConfigRequest(
+          sharingLink: WorkplaceActionConfigRequest.fromEntity(config.addAsLink),
+          downloadLink: config.addAsAttachment == null
+              ? null
+              : WorkplaceActionConfigRequest.fromEntity(config.addAsAttachment!),
+          theme: WorkplaceThemeConfigRequest.fromEntity(config.theme),
+        ),
+      ),
+    ),
+  ).toJson();
 
   @override
   Future<String> exchangeToken(Uri platformUrl, String oidcIdToken) async {
@@ -59,7 +140,7 @@ class WorkplaceDataSourceImpl implements WorkplaceDataSource {
         exchangeType: WorkplaceExchangeType.app,
       ).toJson(),
     );
-    final data = WorkplaceExchangeTokenResponse.fromJson(asJsonMap(response.data));
+    final data = WorkplaceExchangeTokenResponse.fromJson(_asJsonMap(response.data));
     final accessToken = data.accessToken;
     return accessToken;
   }
