+++ workplace/lib/domain/usecase/create_drive_intent_interactor.dart
@@ -1,6 +1,7 @@
 import 'package:core/presentation/state/failure.dart';
 import 'package:core/presentation/state/success.dart';
 import 'package:dartz/dartz.dart';
+import '../entity/workplace_intent_access_mode.dart';
 import '../entity/workplace_intent_config.dart';
 import '../repository/workplace_repository.dart';
 import '../state/workplace_intent_state.dart';
@@ -12,14 +13,14 @@ class CreateDriveIntentInteractor {
 
   Stream<Either<Failure, Success>> execute(
     Uri platformUrl,
-    String accessToken, {
+    WorkplaceIntentAccessMode accessMode, {
     required WorkplaceIntentConfig config,
   }) async* {
     try {
       yield Right(CreatingWorkplaceIntent());
       final intent = await _repository.createIntent(
         platformUrl: platformUrl,
-        accessToken: accessToken,
+        accessMode: accessMode,
         config: config,
       );
       yield Right(CreateWorkplaceIntentSuccess(intent));
