+++ workplace/test/data/workplace_datasource_impl_test.dart
@@ -7,6 +7,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:workplace/data/datasource_impl/workplace_datasource_impl.dart';
 import 'package:workplace/data/workplace_dio.dart';
 import 'package:workplace/domain/entity/workplace_action_config.dart';
+import 'package:workplace/domain/entity/workplace_intent_access_mode.dart';
 import 'package:workplace/domain/entity/workplace_intent_config.dart';
 import 'package:workplace/domain/entity/workplace_theme.dart';
 
@@ -59,6 +60,163 @@ void main() {
     datasource = WorkplaceDataSourceImpl();
   });
 
+  Map<String, dynamic> buildResponse({
+    required String id,
+    required String href,
+  }) => {
+    'data': {
+      'id': id,
+      'attributes': {
+        'action': 'PICK',
+        'type': 'files',
+        'permissions': ['GET'],
+        'services': [
+          {'href': href},
+        ],
+      },
+    },
+  };
+
+  group('WorkplaceDataSourceImpl::parseIntentResponse::', () {
+    test('Should return WorkplaceIntent with correct id and url', () {
+      final result = datasource.parseIntentResponse(
+        buildResponse(id: 'intent-1', href: 'https://drive.example.com/pick'),
+        requireHttps: false,
+      );
+
+      expect(result.intentId, equals('intent-1'));
+      expect(result.intentUrl, equals(Uri.parse('https://drive.example.com/pick')));
+    });
+
+    test('Should parse JSON string input', () {
+      const jsonString =
+          '{"data":{"id":"intent-str","attributes":{"action":"PICK","type":"files","permissions":["GET"],"services":[{"href":"https://drive.example.com/str"}]}}}';
+
+      final result = datasource.parseIntentResponse(
+        jsonString,
+        requireHttps: false,
+      );
+
+      expect(result.intentId, equals('intent-str'));
+      expect(result.intentUrl.host, equals('drive.example.com'));
+    });
+
+    test('Should throw StateError when services list is empty', () {
+      final data = {
+        'data': {
+          'id': 'intent-1',
+          'attributes': {
+            'action': 'PICK',
+            'type': 'files',
+            'permissions': ['GET'],
+            'services': <dynamic>[],
+          },
+        },
+      };
+
+      expect(
+        () => datasource.parseIntentResponse(data, requireHttps: false),
+        throwsA(isA<StateError>()),
+      );
+    });
+
+    test('Should throw FormatException for unsupported input type', () {
+      expect(
+        () => datasource.parseIntentResponse(42, requireHttps: false),
+        throwsA(isA<FormatException>()),
+      );
+    });
+
+    test('Should throw ArgumentError when requireHttps is true and URL is http', () {
+      expect(
+        () => datasource.parseIntentResponse(
+          buildResponse(id: 'intent-1', href: 'http://drive.example.com/pick'),
+          requireHttps: true,
+        ),
+        throwsA(isA<ArgumentError>()),
+      );
+    });
+
+    test('Should accept http URL when requireHttps is false', () {
+      final result = datasource.parseIntentResponse(
+        buildResponse(id: 'intent-1', href: 'http://drive.example.com/pick'),
+        requireHttps: false,
+      );
+
+      expect(result.intentUrl.scheme, equals('http'));
+    });
+
+    test('Should use kReleaseMode as default for requireHttps', () {
+      // In debug/test mode, kReleaseMode is false so http URLs are accepted
+      if (kReleaseMode) {
+        expect(
+          () => datasource.parseIntentResponse(
+            buildResponse(id: 'intent-1', href: 'http://drive.example.com/pick'),
+          ),
+          throwsA(isA<ArgumentError>()),
+        );
+      } else {
+        expect(
+          datasource.parseIntentResponse(
+            buildResponse(id: 'intent-1', href: 'http://drive.example.com/pick'),
+          ),
+          isNotNull,
+        );
+      }
+    });
+
+    test('Should parse client from attributes when present', () {
+      final data = {
+        'data': {
+          'id': 'intent-1',
+          'attributes': {
+            'action': 'PICK',
+            'type': 'files',
+            'permissions': ['GET'],
+            'services': [
+              {'href': 'https://drive.example.com/pick'},
+            ],
+            'client': 'client-abc',
+          },
+        },
+      };
+
+      final result = datasource.parseIntentResponse(data, requireHttps: false);
+
+      expect(result.client, equals('client-abc'));
+    });
+
+    test('Should have null client when attributes omit it', () {
+      final result = datasource.parseIntentResponse(
+        buildResponse(id: 'intent-1', href: 'https://drive.example.com/pick'),
+        requireHttps: false,
+      );
+
+      expect(result.client, isNull);
+    });
+
+    test('Should use first service href when multiple services are present', () {
+      final data = {
+        'data': {
+          'id': 'intent-multi',
+          'attributes': {
+            'action': 'PICK',
+            'type': 'files',
+            'permissions': ['GET'],
+            'services': [
+              {'href': 'https://drive.example.com/first'},
+              {'href': 'https://drive.example.com/second'},
+            ],
+          },
+        },
+      };
+
+      final result = datasource.parseIntentResponse(data, requireHttps: false);
+
+      expect(result.intentUrl, equals(Uri.parse('https://drive.example.com/first')));
+    });
+  });
+
   group('WorkplaceDataSourceImpl::createIntent::', () {
     late Dio originalDio;
 
@@ -90,7 +248,7 @@ void main() {
 
       final result = await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com'),
-        accessToken: 'test-token',
+        accessMode: const BearerTokenAccessMode('test-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           addAsAttachment: WorkplaceActionConfig(label: 'https://attach.url'),
@@ -108,7 +266,7 @@ void main() {
 
       await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com/api/'),
-        accessToken: 'test-token',
+        accessMode: const BearerTokenAccessMode('test-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           addAsAttachment: WorkplaceActionConfig(label: 'https://attach.url'),
@@ -128,7 +286,7 @@ void main() {
 
       await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com'),
-        accessToken: 'test-token',
+        accessMode: const BearerTokenAccessMode('test-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           addAsAttachment: WorkplaceActionConfig(label: 'https://attach.url'),
@@ -148,7 +306,7 @@ void main() {
 
       await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com'),
-        accessToken: 'my-secret-token',
+        accessMode: const BearerTokenAccessMode('my-secret-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           addAsAttachment: WorkplaceActionConfig(label: 'https://attach.url'),
@@ -168,7 +326,7 @@ void main() {
 
       await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com'),
-        accessToken: 'test-token',
+        accessMode: const BearerTokenAccessMode('test-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           addAsAttachment: WorkplaceActionConfig(label: 'https://attach.url'),
@@ -201,7 +359,7 @@ void main() {
 
       await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com'),
-        accessToken: 'test-token',
+        accessMode: const BearerTokenAccessMode('test-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           addAsAttachment: WorkplaceActionConfig(
@@ -229,7 +387,7 @@ void main() {
 
       await datasource.createIntent(
         platformUrl: Uri.parse('https://platform.example.com'),
-        accessToken: 'test-token',
+        accessMode: const BearerTokenAccessMode('test-token'),
         config: const WorkplaceIntentConfig(
           addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
           theme: WorkplaceTheme.light,
@@ -250,7 +408,7 @@ void main() {
       expect(
         () => datasource.createIntent(
           platformUrl: Uri.parse('https://platform.example.com'),
-          accessToken: 'test-token',
+          accessMode: const BearerTokenAccessMode('test-token'),
           config: const WorkplaceIntentConfig(
             addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
             addAsAttachment: WorkplaceActionConfig(label: 'https://attach.url'),
@@ -268,7 +426,7 @@ void main() {
       expect(
         () => datasource.createIntent(
           platformUrl: Uri.parse('https://platform.example.com'),
-          accessToken: '',
+          accessMode: const BearerTokenAccessMode(''),
           config: const WorkplaceIntentConfig(
             addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
             theme: WorkplaceTheme.light,
@@ -285,7 +443,7 @@ void main() {
       expect(
         () => datasource.createIntent(
           platformUrl: Uri.parse('https://platform.example.com'),
-          accessToken: '   ',
+          accessMode: const BearerTokenAccessMode('   '),
           config: const WorkplaceIntentConfig(
             addAsLink: WorkplaceActionConfig(label: 'https://link.url'),
             theme: WorkplaceTheme.light,
